Finished HW3
