+++ HW3/app/src/main/java/com/example/hw3/MainActivity.java
@@ -1,49 +1,13 @@
 package com.example.hw3;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.NotificationCompat;
-import androidx.core.app.NotificationManagerCompat;
-
-import android.app.AlarmManager;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
-import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
 
-import java.util.Calendar;
-
-import static com.example.hw3.App.CHANNEL_ID;
-
 public class MainActivity extends AppCompatActivity {
-
-    private AlarmManager alarmManager;
-    private PendingIntent alarmIntent;
-    //private NotificationManagerCompat notificationManager;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        startAlarm(this);
-        /**
-        notificationManager = NotificationManagerCompat.from(this);
-        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, CHANNEL_ID)
-                .setSmallIcon(R.mipmap.ic_launcher)
-                .setContentTitle("Quote")
-                .setContentText("123");
-
-        notificationManager.notify(1, builder.build());*/
-
-    }
-
-    private void startAlarm(Context context){
-        alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
-        Intent intent = new Intent(context, AlarmReceiver.class);
-
-        alarmIntent = PendingIntent.getBroadcast(context, 1, intent, 0);
-
-        alarmManager.setRepeating(AlarmManager.RTC_WAKEUP, System.currentTimeMillis(),
-                                        5*1000, alarmIntent);
+        QuoteIntentService.startActionQuote(this);
     }
 }
